Adds the support_functions/ directory dependence to makeTables/ instead of having to call functions from personal home directories
